Refactored code to meet best practises for naming convention
