--- backend/SectorService/src/main/java/com/sankalp/SectorService/Entities/Sector.java
@@ -1,5 +1,0 @@
-package com.sankalp.SectorService.Entities;
-
-public class Sector {
-
-}
