+++ backend/exchange-service/pom.xml
@@ -4,14 +4,14 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.springframework</groupId>
-		<artifactId>Stock-Market-Application</artifactId>
+		<artifactId>stock-market-application</artifactId>
 		<version>0.1.0</version>
 		<relativePath>../</relativePath>
 	</parent>
 	<groupId>com.sankalp</groupId>
-	<artifactId>ExchangeService</artifactId>
+	<artifactId>exchange-service</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>ExchangeService</name>
+	<name>exchange-service</name>
 	<description>Exchange Microservice for Stock Market Application</description>
 	<properties>
 		<java.version>11</java.version>
