+++ backend/exchange-service/src/main/java/com/sankalp/exchangeservice/entity/Address.java
@@ -1,4 +1,4 @@
-package com.sankalp.ExchangeService.Entities;
+package com.sankalp.exchangeservice.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
