+++ backend/exchange-service/src/main/java/com/sankalp/exchangeservice/repository/StockExchangeRepository.java
@@ -1,9 +1,9 @@
-package com.sankalp.ExchangeService.Repositories;
+package com.sankalp.exchangeservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sankalp.ExchangeService.Entities.StockExchange;
+import com.sankalp.exchangeservice.entity.StockExchange;
 
 @Repository
 public interface StockExchangeRepository extends JpaRepository<StockExchange, Integer> {
