+++ backend/sector-service/pom.xml
@@ -4,14 +4,14 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.springframework</groupId>
-		<artifactId>Stock-Market-Application</artifactId>
+		<artifactId>stock-market-application</artifactId>
 		<version>0.1.0</version>
 		<relativePath>../</relativePath>
 	</parent>
 	<groupId>com.sankalp</groupId>
-	<artifactId>SectorService</artifactId>
+	<artifactId>sector-service</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>SectorService</name>
+	<name>sector-service</name>
 	<description>Microservice for Sector</description>
 	<properties>
 		<java.version>11</java.version>
